feat!: redesign CLI surface for simplicity and predictability
Reduce command count from 9 to 7 with clearer naming:
- Merge `list` + `print` into `show` (with --paths, --reveal, --format)
- Replace `up` (reconciler) with read-only `validate` command
- Rename `env:create` to `env:add`

Key improvements:
- `show` without --reveal never contacts the provider (no credentials needed)
- `show --format json` always outputs consistent tree shape (no more split {config,secrets})
- Secrets display as `<secret>` placeholder when not revealing
- `validate` uses provider.list() instead of try/catch on get()

Removes the entire reconciler subsystem (diff, apply, input, render, types)
except topo-sort which validate still uses.

BREAKING CHANGE: removes `list`, `print`, `up`, `env:create` commands.
Use `show --paths`, `show`, `validate`, `env:add` respectively.
