--- test/core/reconciler/input.test.ts
@@ -1,93 +1,0 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { makeCollector, readMaskedLine } from '../../../src/core/reconciler/input.js';
-
-describe('readMaskedLine', () => {
-    it('resolves correctly when a multi-character chunk with newline is delivered at once', async () => {
-        const mockStdin = {
-            setRawMode: vi.fn(),
-            on: vi.fn(),
-            removeListener: vi.fn(),
-            resume: vi.fn(),
-            pause: vi.fn()
-        };
-        const mockStdout = { write: vi.fn() };
-
-        const originalStdin = process.stdin;
-        const originalStdout = process.stdout;
-        Object.defineProperty(process, 'stdin', { value: mockStdin, configurable: true });
-        Object.defineProperty(process, 'stdout', { value: mockStdout, configurable: true });
-
-        let capturedHandler: ((chunk: Buffer) => void) | undefined;
-        mockStdin.on.mockImplementation((_event: string, handler: (chunk: Buffer) => void) => {
-            capturedHandler = handler;
-        });
-
-        const resultPromise = readMaskedLine('Enter secret: ');
-
-        // Simulate a multi-character chunk delivered at once (e.g. pasted input)
-        capturedHandler!(Buffer.from('secret\n'));
-
-        const result = await resultPromise;
-        expect(result).toBe('secret');
-
-        Object.defineProperty(process, 'stdin', { value: originalStdin, configurable: true });
-        Object.defineProperty(process, 'stdout', { value: originalStdout, configurable: true });
-    });
-});
-
-describe('makeCollector', () => {
-    describe('prompt source', () => {
-        it('reads value from stdin with a prompt derived from the secret path', async () => {
-            const mockStdin = {
-                setRawMode: vi.fn(),
-                on: vi.fn(),
-                removeListener: vi.fn(),
-                resume: vi.fn(),
-                pause: vi.fn()
-            };
-            const mockStdout = { write: vi.fn() };
-
-            const originalStdin = process.stdin;
-            const originalStdout = process.stdout;
-            Object.defineProperty(process, 'stdin', { value: mockStdin, configurable: true });
-            Object.defineProperty(process, 'stdout', { value: mockStdout, configurable: true });
-
-            let capturedHandler: ((chunk: Buffer) => void) | undefined;
-            mockStdin.on.mockImplementation((_event: string, handler: (chunk: Buffer) => void) => {
-                capturedHandler = handler;
-            });
-
-            const collect = makeCollector({ kind: 'prompt' });
-            const resultPromise = collect('api/token');
-
-            capturedHandler!(Buffer.from('my-secret\n'));
-
-            const result = await resultPromise;
-            expect(result).toBe('my-secret');
-
-            const promptArg = (mockStdout.write as ReturnType<typeof vi.fn>).mock.calls[0][0];
-            expect(promptArg).toContain('api/token');
-
-            Object.defineProperty(process, 'stdin', { value: originalStdin, configurable: true });
-            Object.defineProperty(process, 'stdout', { value: originalStdout, configurable: true });
-        });
-    });
-
-    describe('file source', () => {
-        it('returns value from file values for matching path', async () => {
-            const collect = makeCollector({
-                kind: 'file',
-                values: { 'db/password': 'secret123' }
-            });
-
-            const value = await collect('db/password');
-            expect(value).toBe('secret123');
-        });
-
-        it('throws when path is not in file', async () => {
-            const collect = makeCollector({ kind: 'file', values: {} });
-
-            await expect(collect('missing/path')).rejects.toThrow(/missing\/path/);
-        });
-    });
-});
